+++ packages/core/types/src/http/inventory/admin/responses.ts
@@ -6,10 +6,16 @@ import {
 import { AdminInventoryItem } from "./entities"
 
 export interface AdminInventoryItemResponse {
+  /**
+   * The inventory item's details.
+   */
   inventory_item: AdminInventoryItem
 }
 
 export type AdminInventoryItemListResponse = PaginatedResponse<{
+  /**
+   * The list of inventory items.
+   */
   inventory_items: AdminInventoryItem[]
 }>