+++ packages/medusa/src/api/store/products/validators.ts
@@ -1,6 +1,9 @@
 import { z } from "zod"
 import {
+  applyAndAndOrOperators,
   GetProductsParams,
+  recursivelyNormalizeSchema,
+  StoreGetProductParamsDirectFields,
   transformProductParams,
 } from "../../utils/common-validators"
 import {
@@ -9,17 +12,27 @@ import {
   createSelectParams,
 } from "../../utils/validators"
 
+export const StoreGetProductParamsFields = z.object({
+  region_id: z.string().optional(),
+  country_code: z.string().optional(),
+  province: z.string().optional(),
+  cart_id: z.string().optional(),
+})
+
 export type StoreGetProductParamsType = z.infer<typeof StoreGetProductParams>
 
 export const StoreGetProductParams = createSelectParams().merge(
-  // These are used to populate the tax and pricing context
-  z.object({
-    region_id: z.string().optional(),
-    country_code: z.string().optional(),
-    province: z.string().optional(),
-    cart_id: z.string().optional(),
-  })
+  StoreGetProductParamsFields
 )
+
+export const StoreGetProductVariantsParamsFields = z.object({
+  q: z.string().optional(),
+  id: z.union([z.string(), z.array(z.string())]).optional(),
+  options: z.object({ value: z.string(), option_id: z.string() }).optional(),
+  created_at: createOperatorMap().optional(),
+  updated_at: createOperatorMap().optional(),
+  deleted_at: createOperatorMap().optional(),
+})
 
 export type StoreGetProductVariantsParamsType = z.infer<
   typeof StoreGetProductVariantsParams
@@ -27,46 +40,40 @@ export type StoreGetProductVariantsParamsType = z.infer<
 export const StoreGetProductVariantsParams = createFindParams({
   offset: 0,
   limit: 50,
-}).merge(
-  z.object({
-    q: z.string().optional(),
-    id: z.union([z.string(), z.array(z.string())]).optional(),
-    options: z.object({ value: z.string(), option_id: z.string() }).optional(),
-    created_at: createOperatorMap().optional(),
-    updated_at: createOperatorMap().optional(),
-    deleted_at: createOperatorMap().optional(),
-    $and: z.lazy(() => StoreGetProductsParams.array()).optional(),
-    $or: z.lazy(() => StoreGetProductsParams.array()).optional(),
+})
+  .merge(StoreGetProductVariantsParamsFields)
+  .merge(applyAndAndOrOperators(StoreGetProductVariantsParamsFields))
+
+export const StoreGetProductsParamsFields = z
+  .object({
+    region_id: z.string().optional(),
+    country_code: z.string().optional(),
+    province: z.string().optional(),
+    cart_id: z.string().optional(),
   })
-)
+  .merge(GetProductsParams)
+  .strict()
 
 export type StoreGetProductsParamsType = z.infer<typeof StoreGetProductsParams>
 export const StoreGetProductsParams = createFindParams({
   offset: 0,
   limit: 50,
 })
+  .merge(StoreGetProductsParamsFields)
   .merge(
     z
       .object({
-        // These are used to populate the tax and pricing context
-        region_id: z.string().optional(),
-        country_code: z.string().optional(),
-        province: z.string().optional(),
-        cart_id: z.string().optional(),
-
         variants: z
+
           .object({
             options: z
               .object({ value: z.string(), option_id: z.string() })
               .optional(),
-            $and: z.lazy(() => StoreGetProductsParams.array()).optional(),
-            $or: z.lazy(() => StoreGetProductsParams.array()).optional(),
           })
+          .merge(applyAndAndOrOperators(StoreGetProductVariantsParamsFields))
           .optional(),
-        $and: z.lazy(() => StoreGetProductsParams.array()).optional(),
-        $or: z.lazy(() => StoreGetProductsParams.array()).optional(),
       })
-      .merge(GetProductsParams)
+      .merge(applyAndAndOrOperators(StoreGetProductParamsDirectFields))
       .strict()
   )
-  .transform(transformProductParams)
+  .transform(recursivelyNormalizeSchema(transformProductParams))