+++ www/packages/docs-ui/src/components/MainNav/DesktopMenu/index.tsx
@@ -8,6 +8,7 @@ import {
 } from "@medusajs/icons"
 import React, { useMemo, useRef, useState } from "react"
 import {
+  BookIcon,
   Button,
   getOsShortcut,
   Menu,
@@ -35,8 +36,14 @@ export const MainNavDesktopMenu = () => {
         type: "link",
         icon: <HouseIcon />,
         title: "Homepage",
-        link: "https://docs.medusajs.com/v2",
+        link: "https://medusajs.com",
       },
+      // {
+      //   type: "link",
+      //   icon: <BookIcon />,
+      //   title: "Medusa v1",
+      //   link: "https://docs.medusajs.com",
+      // },
       {
         type: "link",
         icon: <TimelineVertical />,